Follow PEPs 8 and 257, and some code-tidying
I made some changes which may not be loved; I'll outline those here for review:

There is a breaking change in this PR, as startdate and enddate are now
called trim_start and trim_end: I renamed these in the code to reflect this.

I also removed much of the explicit type-checking (in the name of ducktyping).
Additionally, I removed the checked for allowable param values, as those are
documented and explicitely checking them only adds overheads and, if at some
point Quandl allows other options, would require changes to this library to
implement.